refactor: sdk with paths, change integrations to client and another tips
