+++ packages/sdk/src/client.ts
@@ -1,1 +1,3 @@
 "use client";
+
+export const check = () => "client";
