+++ packages/sdk/src/index.ts
@@ -1,11 +1,11 @@
-import { ShopifyClient } from "./integrations/shopify";
-import { appRouter } from "./trpc/root";
-import { createTRPCContext } from "./trpc/trpc";
+import { ShopifyClient } from "@sdk/integrations/clients/shopify";
+import { appRouter } from "@sdk/trpc/root";
+import { createTRPCContext } from "@sdk/trpc/trpc";
 
-const integrations = {
+const clients = {
   shopify: {
     client: ShopifyClient,
   },
 };
 
-export { appRouter, createTRPCContext, integrations };
+export { appRouter, clients, createTRPCContext };
