+++ packages/sdk/src/integrations/clients/shopify/fragments/image.ts
