+++ packages/sdk/src/integrations/clients/shopify/fragments/product.ts
