+++ packages/sdk/src/utils/reshapes/edges.ts
@@ -1,4 +1,4 @@
-import { Connection } from '../types';
+import { Connection } from "@sdk/utils/types";
 
 export const removeEdgesAndNodes = <T>(array: Connection<T>): T[] => {
   return array.edges.map((edge) => edge?.node);
