+++ packages/sdk/tsconfig.json
@@ -1,6 +1,11 @@
 {
   "extends": "tsconfig/base.json",
-  "compilerOptions": {},
+  "compilerOptions": {
+    "baseUrl": ".",
+    "paths": {
+      "@sdk/*": ["src/*"]
+    }
+  },
   "include": ["."],
   "exclude": ["node_modules"]
 }
