+++ packages/sdk/tsup.config.ts
@@ -3,7 +3,7 @@ import { defineConfig, Options } from "tsup";
 export const tsup = defineConfig((options: Options) => ({
   name: "server",
   entry: ["src/index.ts", "src/client.ts"],
-  format: ["esm", "cjs"],
+  format: ["esm"],
   outDir: "dist",
   dts: true,
   sourcemap: true,
